Refactoring of the rendering stuff.
