+++ textworld/render/__init__.py
@@ -1,4 +1,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT license.
 
+
+from textworld.render.render import WebdriverNotFoundError
+from textworld.render.render import get_webdriver
 from textworld.render.render import load_state, load_state_from_game_state, visualize
